% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bodi.-package.R
\docType{package}
\name{Bodi-package}
\alias{Bodi-package}
\alias{bodi-package}
\alias{bodi}
\title{Bodi: Boosting Diversity Algorithm}
\description{
We provide an implementation of the boosting diversity algorithm. This is a 
gradient boosting-based algorithm by incorporating a diversity term to
guide the gradient boosting iterations. The idea is to trade off some individual 
optimality for global enhancement. The improvement is obtained with progressively 
generated predictors by boosting diversity. See Borel et al. (2021) <https://hal.archives-ouvertes.fr/hal-03041309v1>
}
\references{
\itemize{ \item Mathias Bourel, Jairo Cugliari, Yannig Goude, Jean-Michel Poggi. Boosting Diversity in Regression Ensembles. 
https://hal.archives-ouvertes.fr/hal-03041309v1 (2021). }
}
\author{
Yannig Goude [aut, cre], Mathias Bourel [aut], Jairo Cugliari [aut], Jean-Michel Poggi [aut]\cr

Mantainer: Yannig Goude <yannig.goude@edf.fr>
}
\keyword{Boosting}
\keyword{Ensemble}
\keyword{Regression}
