% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBlock.R
\name{buildBlock}
\alias{buildBlock}
\title{buildBlock}
\usage{
buildBlock(Nblock, data0)
}
\arguments{
\item{Nblock}{number of blocks}

\item{data0}{the learning set}
}
\value{
A list of vectors containing the indices of each block.
}
\description{
Compute blocks of consecutive data for blockwise CV or sampling.
}
\examples{
buildBlock(4, data.frame(id = 1:15))
}
\author{
Yannig Goude <yannig.goude@edf.fr>
}
